Refactor and add KDoc via ChatGPT
